make allocator adapt multi tables

+++ src/tuple.rs
@@ -3,12 +3,12 @@ pub mod tuple {
 
     #[derive(Clone)]
     pub struct Tuple {
-        pub internal_id: usize,
+        pub internal_id: u64,
         pub fields: Vec<Field>,
     }
 
     impl Tuple {
-        pub fn new(internal_id: usize, fields: Vec<Field>) -> Tuple {
+        pub fn new(internal_id: u64, fields: Vec<Field>) -> Tuple {
             Tuple {
                 internal_id: internal_id,
                 fields: fields,
